requirements.txt
python_telegram_bot==12.5.1
selenium==3.141.0
requests
